# rgw/putobj: offload md5 calculations to the global execution context