--- a/src/rgw/CMakeLists.txt
+++ b/src/rgw/CMakeLists.txt
@@ -81,6 +81,7 @@ set(librgw_common_srcs
   rgw_ldap.cc
   rgw_lc.cc
   rgw_lc_s3.cc
+  rgw_md5.cc
   rgw_metadata.cc
   rgw_multi.cc
   rgw_multi_del.cc