--- a/src/rgw/rgw_md5.h
+++ b/src/rgw/rgw_md5.h
@@ -17,12 +17,18 @@
 
 #include <memory>
 #include <string>
+#include <boost/asio/any_io_executor.hpp>
 #include "rgw_sal_fwd.h"
+
+class optional_yield;
 
 namespace rgw {
 
 /// Create a DataProcessor filter that calculates the md5sum of processed data
 auto create_md5_putobj_pipe(sal::DataProcessor* next,
+                            optional_yield y,
+                            boost::asio::any_io_executor hash_executor,
+                            size_t window_size,
                             std::string& output)
     -> std::unique_ptr<sal::DataProcessor>;