--- a/src/rgw/rgw_putobj_hash.h
+++ b/src/rgw/rgw_putobj_hash.h
@@ -18,6 +18,16 @@
 #include <algorithm>
 #include <string>
 
+#include <boost/asio/any_completion_handler.hpp>
+#include <boost/asio/async_result.hpp>
+#include <boost/asio/bind_executor.hpp>
+#include <boost/asio/dispatch.hpp>
+#include <boost/asio/execution/outstanding_work.hpp>
+#include <boost/asio/post.hpp>
+#include <boost/asio/prefer.hpp>
+
+#include "common/async/yield_context.h"
+
 #include "rgw_putobj.h"
 
 namespace rgw::putobj {
@@ -71,6 +81,144 @@ class HashPipe : public putobj::Pipe {
     }
 
     return Pipe::process(std::move(data), logical_offset);
+  }
+};
+
+/// A streaming data processor that offloads the hashing of incoming bytes to
+/// another executor while forwarding them to the wrapped processor.
+///
+/// Because the hashing is asynchronous with respect to process(), the data from
+/// later process() calls may need to be buffered until the previous hashes
+/// complete. The total amount of buffered memory is limited to the given
+/// window_size.
+///
+/// When the processor is flushed by calling process() with an empty buffer,
+/// the final sum is copied to the given output string.
+///
+/// This class is not thread-safe. All members functions must be called from
+/// within the yield_context's strand executor.
+template <typename Digest>
+class AsyncHashPipe : public putobj::Pipe {
+ public:
+  template <typename ...DigestArgs>
+  AsyncHashPipe(sal::DataProcessor* next,
+                spawn::yield_context yield,
+                boost::asio::any_io_executor hash_executor,
+                size_t window_size,
+                std::string& output,
+                DigestArgs&& ...args)
+    : Pipe(next), yield(yield), hash_executor(hash_executor),
+      window_size(window_size), output(output),
+      digest(std::forward<DigestArgs>(args)...)
+  {}
+
+  ~AsyncHashPipe()
+  {
+    // stop submitting new hash requests and await pending completion
+    pending.clear();
+    if (outstanding) {
+      async_wait(yield);
+    }
+  }
+
+  int process(bufferlist&& data, uint64_t logical_offset) override
+  {
+    if (data.length() == 0) {
+      // wait for all pending hashes to complete
+      while (outstanding) {
+        async_wait(yield);
+      }
+
+      // finalize the digest
+      output = finalize(digest);
+
+      return Pipe::process(std::move(data), logical_offset);
+    }
+
+    // block if the window is full
+    while (pending.length() + outstanding > 0 &&
+           pending.length() + outstanding + data.length() > window_size) {
+      async_wait(yield);
+    }
+    pending.append(data);
+
+    if (!outstanding) {
+      submit(std::move(pending));
+    }
+
+    return Pipe::process(std::move(data), logical_offset);
+  }
+
+ private:
+  spawn::yield_context yield;
+  boost::asio::any_io_executor hash_executor;
+  const size_t window_size;
+  std::string& output;
+  Digest digest;
+
+  // list of buffer segments pending submission
+  bufferlist pending;
+  // number of submitted bytes pending completion
+  size_t outstanding = 0;
+  // completion handler for async_wait(), if any
+  boost::asio::any_completion_handler<void()> waiter;
+
+  // return the coroutine's associated strand executor
+  static auto get_strand_executor(spawn::yield_context yield)
+  {
+    boost::asio::async_completion<spawn::yield_context, void()> init(yield);
+    return get_associated_executor(init.completion_handler);
+  }
+
+  // submit buffers for async hashing
+  void submit(bufferlist&& bl)
+  {
+    ceph_assert(!outstanding);
+    outstanding = bl.length();
+
+    // bind a completion handler to the coroutine's strand executor
+    auto handler = boost::asio::bind_executor(
+        boost::asio::prefer(get_strand_executor(yield),
+                            boost::asio::execution::outstanding_work.tracked),
+        [this] { complete(); });
+
+    // offload the hashing to the requested hash_executor
+    boost::asio::post(hash_executor,
+        [&d = digest, bl = std::move(bl), h = std::move(handler)] () mutable {
+          // hash each buffer segment
+          for (const auto& ptr : bl.buffers()) {
+            d.Update(reinterpret_cast<const unsigned char*>(ptr.c_str()),
+                     ptr.length());
+          }
+          // dispatch the completion back to its executor
+          boost::asio::dispatch(std::move(h));
+        });
+  }
+
+  // hash completion callback
+  void complete()
+  {
+    ceph_assert(outstanding);
+    outstanding = 0;
+
+    if (pending.length()) {
+      // submit pending buffers
+      submit(std::move(pending));
+    }
+
+    if (waiter) { // wake if blocked
+      boost::asio::dispatch(std::move(waiter));
+    }
+  }
+
+  // capture the completion handler in `waiter`
+  template <typename CompletionToken>
+  auto async_wait(CompletionToken&& token)
+  {
+    return boost::asio::async_initiate<CompletionToken, void()>(
+        [this] (auto&& handler) {
+          waiter = std::move(handler);
+        }, token);
   }
 };