* uses the `sal::DataProcessor` interface to move etag calculation logic out of `RGWPutObj::execute()`
* adds an `Async` implementation that offloads etag calculations to another thread (the global thread pool executor) so they can run in parallel with `RGWPutObj::execute()`'s reads from the frontend socket and writes to rados

this offloading is conceptually similar to prior work in https://github.com/ceph/ceph/pull/52920, but without spawning a new thread for each request

this also relates to ongoing work in https://github.com/ceph/ceph/pull/52385 and https://github.com/ceph/ceph/pull/52488 to take advantage of vectorization (and potentially hardware acceleration) for etag calculation. however, that strategy involves a tradeoff between latency and batching, so is targetted at scaled-up workloads to reduce cpu usage. at smaller scales, we may prefer this PR's strategy to improve latency instead

depends on some asio-related commits from https://github.com/ceph/ceph/pull/50055 and https://github.com/ceph/ceph/pull/52496

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
- `jenkins test rook e2e`
</details>