@cbodley 2023-12-31
some preliminary numbers from warp against a vstart cluster on nvme

----
writing 4k objects at concurrency=128:
> $ ./warp put --host localhost:8000 --access-key 0555b35654ad1656d804 --secret-key h7GhxuBLTrlhVUyxSPUKUV8r/2EI4ngqJxD7iBdBYLhwluN30JaT3Q== --concurrent 128 --md5 --obj.size 4096 --autoterm

| 4k PUT | HashPipe  | AsyncHashPipe |
|-| ------------- | ------------- |
| Average | 4.21 MiB/s, 1078.69 obj/s | 4.18 MiB/s, 1069.34 obj/s |
| Fastest | 4.6MiB/s, 1178.52 obj/s | 5.0MiB/s, 1284.92 obj/s |
| 50% Median | 4.1MiB/s, 1059.90 obj/s | 4.2MiB/s, 1075.64 obj/s |
| Slowest | 3.8MiB/s, 975.82 obj/s | 3.4MiB/s, 882.04 obj/s |

----
writing 32m objects at concurrency=128:
> $ ./warp put --host localhost:8000 --access-key 0555b35654ad1656d804 --secret-key h7GhxuBLTrlhVUyxSPUKUV8r/2EI4ngqJxD7iBdBYLhwluN30JaT3Q== --concurrent 128 --md5 --obj.size 33554432 --autoterm

| 32m PUT | HashPipe  | AsyncHashPipe |
|-| ------------- | ------------- |
| Average | 449.77 MiB/s, 14.06 obj/s | 461.63 MiB/s, 14.43 obj/s |
| Fastest | 479.9MiB/s, 15.00 obj/s | 510.9MiB/s, 15.97 obj/s |
| 50% Median | 448.2MiB/s, 14.01 obj/s | 463.0MiB/s, 14.47 obj/s |
| Slowest | 417.9MiB/s, 13.06 obj/s | 430.2MiB/s, 13.44 obj/s |