@mkogan1 2024-01-01
> that's surprising; maybe @mkogan1 can push it around

interesting, will measure CPU utilization with and without and will ramp up the workload, 
possibly the performance difference wont begin to manifest until a CPU constraint is hit