@mkogan1 2024-01-18
expanding on:
| some preliminary numbers from warp against a vstart cluster on nvme

detailing results of additional testing:

Warps behavior compared to Hsbench is sub-optimal when working on small objects because Warp is working on a single bucket - for comparison in screenshot below is radosgw process CPU utilization over PUT workload of 4KB objs
note in the graph below, on left Hsbench which is writing to 80 buckets keeps a constant CPU utilization (and PUT's per second rate) VS Warp on the right which is writing to 1 bucket, while the workload is productivity (radosgw CPU utilization and PUT ops per second) is lowering with bucket aging because of 'rocksdb' cluster aging effect (100% I/O utilization on the rocksdb backing storage device)
hence following tests will be on the larger objects to mitigate this Warp shortcoming
![image](https://github.com/ceph/ceph/assets/31659604/24514001-01cf-421b-b5f3-aae41370777d)



testing with `--obj.size=32M` and load level of `--concurrent 128` and each iteration to a different bucket:

`nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent 128 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear`

this PR shows little slower throughput VS _main_ **and** the CPU utilization of the radosgw process correspondingly lower
![image](https://github.com/ceph/ceph/assets/31659604/ac7bb067-abab-465b-a4ad-604d836b27f6)


but increasing the drive to `--concurrent=250` from `128` flips the results - the radosgw CPU utilization and throughput with this PR are slightly faster than main
`nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear`
![image](https://github.com/ceph/ceph/assets/31659604/e07541a0-bab7-4f97-a910-846c057b63e2)


raw warp data collection numbers of the 32MB tests:
```
// main without PR #55029:
vstart.sh ...

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1245.32 MiB/s, 40.81 obj/s
Throughput, split into 102 x 1s:
 * Fastest:1532.1MiB/s, 50.20 obj/s
 * 50% Median:1240.5MiB/s, 40.65 obj/s
 * Slowest:998.3MiB/s, 32.71 obj/s

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1161.02 MiB/s, 38.04 obj/s
Throughput, split into 105 x 1s:
 * Fastest:1330.6MiB/s, 43.60 obj/s
 * 50% Median:1152.4MiB/s, 37.76 obj/s
 * Slowest:1014.8MiB/s, 33.25 obj/s

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1169.47 MiB/s, 38.32 obj/s
Throughput, split into 107 x 1s:
 * Fastest:1357.3MiB/s, 44.48 obj/s
 * 50% Median:1153.6MiB/s, 37.80 obj/s
 * Slowest:1083.6MiB/s, 35.51 obj/s


// main WITH PR #55029 cherry-pick:
vstart.sh ...

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1224.29 MiB/s, 40.12 obj/s
Throughput, split into 106 x 1s:
 * Fastest:1467.1MiB/s, 48.07 obj/s
 * 50% Median:1239.9MiB/s, 40.63 obj/s
 * Slowest:887.7MiB/s, 29.09 obj/s

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1280.51 MiB/s, 41.96 obj/s
Throughput, split into 105 x 1s:
 * Fastest:1534.7MiB/s, 50.29 obj/s
 * 50% Median:1274.2MiB/s, 41.75 obj/s
 * Slowest:1104.7MiB/s, 36.20 obj/s

nice numactl -N 1 -m 1 -- warp put --host=127.0.0.1:8000 --access-key=b2345678901234567890 --secret-key=b234567890123456789012345678901234567890 --concurrent=250 --obj.size=32M --duration=2m --md5 --bucket="b01b-${RANDOM}" --noclear #--autoterm
----------------------------------------
Operation: PUT. Concurrency: 250
* Average:1224.81 MiB/s, 40.13 obj/s
Throughput, split into 104 x 1s:
 * Fastest:1454.3MiB/s, 47.65 obj/s
 * 50% Median:1209.5MiB/s, 39.63 obj/s
 * Slowest:1124.0MiB/s, 36.83 obj/s
```