@cbodley 2024-01-24
> this also relates to ongoing work in https://github.com/ceph/ceph/pull/52385 and https://github.com/ceph/ceph/pull/52488 to take advantage of vectorization (and potentially hardware acceleration) for etag calculation. however, that strategy involves a tradeoff between latency and batching, so is targetted at scaled-up workloads to reduce cpu usage. at smaller scales, we may prefer this PR's strategy to improve latency instead

given the performance results, this doesn't seem to help at all for smaller workloads